requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.0.0
colorama>=0.4.6
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.14

